requirements.txt
# MutationScan Core Dependencies
# Core bioinformatics and scientific computing
biopython>=1.79,<2.0
numpy>=1.21.0,<2.0
scipy>=1.7.0,<2.0
pandas>=1.3.0,<2.0

# Configuration and file handling
PyYAML>=5.4,<7.0

# Network and API
requests>=2.26.0,<3.0

# Visualization
matplotlib>=3.4.0,<4.0

# Optional: PyMOL for 3D visualization
# Uncomment to install (requires separate MolSoft license)
# pymol-open-source>=2.4.0

# Development dependencies
# Install with: pip install -e ".[dev]"
# pytest>=6.2.5
# pytest-cov>=2.12.1
# black>=21.7b0
# flake8>=3.9.2
# mypy>=0.910
# isort>=5.9.3

# Documentation dependencies
# Install with: pip install -e ".[docs]"
# sphinx>=4.0.0
# sphinx-rtd-theme>=0.5.0
